Rename files to naming convention of project
